+++ src/main/java/mod/bluestaggo/modernerbeta/world/chunk/provider/ChunkProviderEarlyBedrock.java
@@ -1,0 +1,34 @@
+package mod.bluestaggo.modernerbeta.world.chunk.provider;
+
+import mod.bluestaggo.modernerbeta.api.world.spawn.SpawnLocator;
+import mod.bluestaggo.modernerbeta.util.mersenne.MTRandom;
+import mod.bluestaggo.modernerbeta.world.chunk.ModernBetaChunkGenerator;
+import mod.bluestaggo.modernerbeta.world.spawn.SpawnLocatorRelease;
+
+import java.util.Random;
+
+public class ChunkProviderEarlyBedrock extends ChunkProviderMajorRelease {
+    public ChunkProviderEarlyBedrock(ModernBetaChunkGenerator chunkGenerator, long seed) {
+        super(chunkGenerator, seed);
+    }
+
+    @Override
+    protected Random getRandom(long seed) {
+        return new MTRandom(seed);
+    }
+
+    @Override
+    public SpawnLocator getSpawnLocator() {
+        return new SpawnLocatorRelease(this, new MTRandom(this.seed));
+    }
+
+    /*
+     * MCPE uses different values to seed random surface generation.
+     */
+    @Override
+    protected Random createSurfaceRandom(int chunkX, int chunkZ) {
+        long seed = (long)chunkX * 0x14609048 + (long)chunkZ * 0x7ebe2d5;
+
+        return new MTRandom(seed);
+    }
+}