+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## Additions
+- Added Bedrock 1.2 and Bedrock 1.17 presets based on the 1.12.2 and 1.17.1 presets respectively but with Bedrock Edition biome and terrain generation.
+- Added the Early Bedrock chunk provider to allow for the use of Mersenne Twister RNG and the 128 block height limit for forced height worlds.
+
 ## Fixes
 - Fixed plains borders between swamps and deserts, snowy taigas and snowy plains in 1.7+ presets being missing.